[Fix] int to float
공격력 방어력 체력 int에서 float으로 변경

+++ Assets/02.Scripts/Enemy/Data/EnemySO.cs
@@ -5,11 +5,11 @@ using UnityEngine;
 [CreateAssetMenu(fileName = "EnemySO", menuName = "Characters/Enemy")]
 public class EnemySO : ScriptableObject
 {
-    [field: SerializeField] public int Damage { get; private set; }
+    [field: SerializeField] public float Damage { get; private set; }
     [field: SerializeField] public int AttackDelay { get; private set; }
     [field: SerializeField] public float PlayerChasingRange { get; private set; }
     [field: SerializeField] public float AttackRange { get; private set; }
-    [field: SerializeField] public int MaxHealth { get; private set; }
+    [field: SerializeField] public float MaxHealth { get; private set; }
     [field: SerializeField] public int DropEXP { get; private set; }
     [field: SerializeField] public GameObject[] DropItem { get; private set; }
 }
