+++ Assets/02.Scripts/Player/Data/PlayerSkillData.cs
@@ -10,14 +10,14 @@ public class SkillInfoData
     [field: SerializeField] private string SkillName;
     [field: SerializeField] private int SkillCost;
     [field: SerializeField] private float SkillCoolTime;
-    [field: SerializeField] private int SkillDamage;
+    [field: SerializeField] private float SkillDamage;
     [field: SerializeField] private float SkillRange;
     [field: SerializeField] private float SkillDistance;
     [field: SerializeField] private string SkillPrefabsName;
 
     public string GetSkillName() { return SkillName; }
     public int GetSkillCost() { return SkillCost; }
-    public int GetSkillDamage() { return SkillDamage; }
+    public float GetSkillDamage() { return SkillDamage; }
     public float GetSkillCoolTime() { return SkillCoolTime; }
     public float GetSkillRange() { return SkillRange; }
     public float GetSkillDistance() { return SkillDistance; }
