+++ Assets/02.Scripts/UI/Player/StatSlider/PlayerBaseSlider.cs
@@ -7,18 +7,18 @@ public class PlayerBaseSlider : MonoBehaviour
     protected Slider _slider;
     [SerializeField] private TMP_Text _Text;
 
-    [SerializeField] private int _currentValue;
-    [SerializeField] private int _maxValue;
+    [SerializeField] private float _currentValue;
+    [SerializeField] private float _maxValue;
 
     protected virtual void Awake()
     {
         _slider = GetComponent<Slider>();
     }
 
-    protected virtual void ChangeBar(int _currentValue, int _maxValue)
+    protected virtual void ChangeBar(float _currentValue, float _maxValue)
     {
         _Text.text = (_currentValue + "/" + _maxValue);
-        _slider.value = (float)_currentValue / (float)_maxValue;
+        _slider.value =_currentValue / _maxValue;
 
     }
 }
